+++ base/exports.jl
@@ -685,6 +685,7 @@ export
     tril,
     triu!,
     triu,
+    vecdot,
     vecnorm,
     ⋅,
     ×,